refactor(masonry): move @types to /src and update imports
